+++ Tests/RxCocoaTests/UIScrollView+RxTests.swift
@@ -62,6 +62,28 @@ extension UIScrollViewTests {
         XCTAssertTrue(completed)
     }
 	
+    func testScrollViewWillBeginDecelerating() {
+        var completed = false
+
+        autoreleasepool {
+            let scrollView = UIScrollView()
+            var willBeginDecelerating = false
+
+            _ = scrollView.rx.willBeginDecelerating.subscribe(onNext: {
+                willBeginDecelerating = true
+            }, onCompleted: {
+                completed = true
+            })
+
+            XCTAssertFalse(willBeginDecelerating)
+
+            scrollView.delegate!.scrollViewWillBeginDecelerating!(scrollView)
+
+            XCTAssertTrue(willBeginDecelerating)
+        }
+
+        XCTAssertTrue(completed)
+    }
 	
 	func testScrollViewDidEndDecelerating() {
 		var completed = false
@@ -85,7 +107,94 @@ extension UIScrollViewTests {
 		
 		XCTAssertTrue(completed)
 	}
+
+    func testScrollViewWillBeginDragging() {
+        var completed = false
+
+        autoreleasepool {
+            let scrollView = UIScrollView()
+            var willBeginDragging = false
+
+            _ = scrollView.rx.willBeginDragging.subscribe(onNext: {
+                willBeginDragging = true
+            }, onCompleted: {
+                completed = true
+            })
+
+            XCTAssertFalse(willBeginDragging)
+
+            scrollView.delegate!.scrollViewWillBeginDragging!(scrollView)
+
+            XCTAssertTrue(willBeginDragging)
+        }
+
+        XCTAssertTrue(completed)
+    }
 	
+    func testScrollViewWillEndDragging() {
+        var completed = false
+        
+        autoreleasepool {
+            let scrollView = UIScrollView()
+
+            let positiveVelocity = CGPoint(x: 1.5, y: 2.5)
+            var positiveOffset = CGPoint(x: 27.4, y: 853.0)
+
+            let negativeVelocity = CGPoint(x: 1.5, y: 2.5)
+            var zeroOffset = CGPoint.zero
+
+            var velocity: CGPoint? = nil
+            var offset: CGPoint? = nil
+
+            _ = scrollView.rx.willEndDragging.subscribe(onNext: {
+                velocity = $0
+                offset = $1.pointee
+            }, onCompleted: {
+                completed = true
+            })
+            
+            XCTAssertNil(velocity)
+            XCTAssertNil(offset)
+
+            scrollView.delegate!.scrollViewWillEndDragging!(scrollView, withVelocity: positiveVelocity, targetContentOffset: &positiveOffset)
+
+            XCTAssertEqual(positiveVelocity, velocity)
+            XCTAssertEqual(positiveOffset, offset)
+
+            scrollView.delegate!.scrollViewWillEndDragging!(scrollView, withVelocity: negativeVelocity, targetContentOffset: &zeroOffset)
+
+            XCTAssertEqual(negativeVelocity, velocity)
+            XCTAssertEqual(zeroOffset, offset)
+        }
+        
+        XCTAssertTrue(completed)
+    }
+
+    func testScrollViewWillEndDraggingWithModifyingOffset() {
+        var completed = false
+
+        autoreleasepool {
+            let scrollView = UIScrollView()
+
+            var initialOffset = CGPoint(x: 27.4, y: 853.0)
+            let changedOffset = CGPoint(x: 42.5, y: 97.4)
+
+            _ = scrollView.rx.willEndDragging.subscribe(onNext: {
+                $1.pointee = changedOffset
+            }, onCompleted: {
+                completed = true
+            })
+
+            XCTAssertNotEqual(changedOffset, initialOffset)
+
+            scrollView.delegate!.scrollViewWillEndDragging!(scrollView, withVelocity: CGPoint.zero, targetContentOffset: &initialOffset)
+
+            XCTAssertEqual(changedOffset, initialOffset)
+        }
+
+        XCTAssertTrue(completed)
+    }
+
 	func testScrollViewDidEndDragging() {
 		var completed = false
 		
@@ -109,7 +218,7 @@ extension UIScrollViewTests {
 		
 		XCTAssertTrue(completed)
 		
-		}
+    }
 
     func testScrollViewContentOffset() {
         var completed = false
@@ -189,6 +298,62 @@ extension UIScrollViewTests {
             XCTAssertTrue(didEndScrollingAnimation)
         }
         
+        XCTAssertTrue(completed)
+    }
+
+    func testScrollViewWillBeginZooming() {
+        var completed = false
+
+        autoreleasepool {
+            let scrollView = UIScrollView()
+            let zoomView = UIView()
+            var results: [UIView?] = []
+
+            _ = scrollView.rx.willBeginZooming.subscribe(onNext: { value in
+                results.append(value)
+            }, onCompleted: {
+                completed = true
+            })
+
+            XCTAssertTrue(results.isEmpty)
+
+            scrollView.delegate!.scrollViewWillBeginZooming!(scrollView, with: zoomView)
+            scrollView.delegate!.scrollViewWillBeginZooming!(scrollView, with: nil)
+
+            XCTAssertEqual(results[0], zoomView)
+            XCTAssertNil(results[1])
+        }
+
+        XCTAssertTrue(completed)
+    }
+
+    func testScrollViewDidEndZooming() {
+        var completed = false
+
+        autoreleasepool {
+            let scrollView = UIScrollView()
+            let zoomView = UIView()
+            var viewResults: [UIView?] = []
+            var scaleResults: [CGFloat] = []
+
+            _ = scrollView.rx.didEndZooming.subscribe(onNext: { (view, scale) in
+                viewResults.append(view)
+                scaleResults.append(scale)
+            }, onCompleted: {
+                completed = true
+            })
+
+            XCTAssertTrue(viewResults.isEmpty)
+            XCTAssertTrue(scaleResults.isEmpty)
+
+            scrollView.delegate!.scrollViewDidEndZooming!(scrollView, with: zoomView, atScale: 0)
+            scrollView.delegate!.scrollViewDidEndZooming!(scrollView, with: nil, atScale: 2)
+
+            XCTAssertEqual(viewResults[0], zoomView)
+            XCTAssertNil(viewResults[1])
+            XCTAssertEqual(scaleResults, [0, 2])
+        }
+
         XCTAssertTrue(completed)
     }
 }